refactor(extractor): cleaning code
