+++ src/scripts/content/extractor/types/ExtractorPerplexity.js
@@ -1,10 +1,4 @@
 import ExtractorSourcesPerplexity from "../sources/ExtractorSourcesPerplexity";
-import {extractPageCommon} from "../extractPage";
-
-// export async function extractPage (format, metadata) {
-//   console.log("extractPage Perplexity")
-//   return await extractPageCommon(format, metadata, processMessage)
-// }
 
 export async function processMessage(content, format) {
   if (!content.hasChildNodes()) return "";
